Update EmpiresSchematic.java
First argument of DBUpdate is expected to be unique as it is used to store the update in the DB to allow seamless updates without re-running each individual update on server start.

Note: Updates could probably be made to be idempotent and not require them be stored in the DB as it is currently. However, doing like this is easier than attempting to make the operations idempotent.